Add Java code generation support.
Reorganized the protos/ directory structure to be more flat so that the compiler commands are simpler to write, while still maintaining compatibility with previous versions, hopefully.

Added Java generation options.

Added intiial version of the generated Java code.

Added changes to the Go generation that occurred because of the protos/ reorg.